Orquestador_Osint/requirements.txt
dnspython==2.6.1
python-whois==0.9.4
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==10.4.0
reportlab==4.2.5
flask==3.0.3
shodan==1.31.0
networkx==3.2.1
matplotlib==3.8.2